+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- `AttributeIdent` implementation for `Option<impl AttributeIdent>`
+
 ## [0.9.2] - 2024-06-22
 ### Fixed
 - Partial type copies visibility from attribute struct